* Use int32_t for orientation to avoid any loss of accuracy. * Fixed the position calculation. The turning angle was used, instead of the orientation.

+++ lib/LineFollower/LineSensorsCalibrationState.h
@@ -110,7 +110,7 @@ private:
     /**
      * Calibration turn angle in mrad (corresponds to 60°).
      */
-    static const int16_t CALIB_ANGLE = (FP_2PI() / 6);
+    static const int32_t CALIB_ANGLE = (FP_2PI() / 6);
 
     SimpleTimer m_timer;            /**< Timer used to wait, until the calibration drive starts. */
     Phase       m_phase;            /**< Current calibration phase */
